Moved ModuleBundleFactory out of internal package Because it might be needed when implementing, say, Gradle support for a new language
