+++ gradle-spaghetti-plugin/src/main/java/com/prezi/spaghetti/gradle/PackageModule.java
@@ -5,7 +5,7 @@ import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import com.prezi.spaghetti.bundle.ModuleBundle;
 import com.prezi.spaghetti.bundle.internal.ModuleBundleElement;
-import com.prezi.spaghetti.bundle.internal.ModuleBundleFactory;
+import com.prezi.spaghetti.bundle.ModuleBundleFactory;
 import com.prezi.spaghetti.gradle.internal.AbstractBundleModuleTask;
 import com.prezi.spaghetti.packaging.ModulePackageParameters;
 import com.prezi.spaghetti.packaging.ModuleType;
