+++ gradle-spaghetti-plugin/src/main/java/com/prezi/spaghetti/gradle/internal/AbstractBundleModuleTask.java
@@ -5,7 +5,7 @@ import com.google.common.collect.Sets;
 import com.google.common.io.Files;
 import com.prezi.spaghetti.ast.ModuleNode;
 import com.prezi.spaghetti.bundle.ModuleBundle;
-import com.prezi.spaghetti.bundle.internal.ModuleBundleFactory;
+import com.prezi.spaghetti.bundle.ModuleBundleFactory;
 import com.prezi.spaghetti.bundle.internal.ModuleBundleParameters;
 import com.prezi.spaghetti.definition.ModuleConfiguration;
 import com.prezi.spaghetti.generator.JavaScriptBundleProcessor;
