+++ spaghetti-core/src/main/java/com/prezi/spaghetti/bundle/ModuleBundleFactory.java
@@ -1,6 +1,9 @@
-package com.prezi.spaghetti.bundle.internal;
+package com.prezi.spaghetti.bundle;
 
-import com.prezi.spaghetti.bundle.ModuleBundle;
+import com.prezi.spaghetti.bundle.internal.DefaultModuleBundle;
+import com.prezi.spaghetti.bundle.internal.ModuleBundleElement;
+import com.prezi.spaghetti.bundle.internal.ModuleBundleInternal;
+import com.prezi.spaghetti.bundle.internal.ModuleBundleParameters;
 import com.prezi.spaghetti.structure.OutputType;
 import com.prezi.spaghetti.structure.internal.StructuredDirectoryProcessor;
 import com.prezi.spaghetti.structure.internal.StructuredDirectoryWriter;
